+++ src/ImGuiGlfw.Sample/Services/Ui/InputWindow.cs
@@ -1,12 +1,14 @@
 ﻿using ImGuiGlfw.Sample.Utils;
 using ImGuiNET;
 using Silk.NET.GLFW;
+using System.Numerics;
 
 namespace ImGuiGlfw.Sample.Services.Ui;
 
 public sealed class InputWindow
 {
 	private readonly Dictionary<Keys, string> _keyDisplayStringCache = [];
+	private readonly Dictionary<MouseButton, string> _mouseButtonDisplayStringCache = [];
 
 	private readonly GlfwInput _glfwInput;
 
@@ -61,12 +63,65 @@ public sealed class InputWindow
 			ImGui.SeparatorText("GLFW pressed chars");
 
 			ImGui.Text(Inline.Span($"{_glfwInput.CharsPressed.Count} key(s):"));
-			ImGui.SameLine();
-			for (int i = 0; i < _glfwInput.CharsPressed.Count; i++)
+			if (ImGui.BeginChild("CharsPressedChildWindow", new Vector2(0, 48), ImGuiChildFlags.Border, ImGuiWindowFlags.AlwaysVerticalScrollbar))
 			{
-				ImGui.Text(Inline.Span((char)_glfwInput.CharsPressed[i]));
-				ImGui.SameLine();
+				for (int i = 0; i < _glfwInput.CharsPressed.Count; i++)
+				{
+					ImGui.Text(Inline.Span((char)_glfwInput.CharsPressed[i]));
+				}
 			}
+
+			ImGui.EndChild();
+
+			ImGui.SeparatorText("Space key state");
+
+			ImGui.Text(Inline.Span($"Down: {(_glfwInput.IsKeyDown(Keys.Space) ? "true" : "false")}"));
+			ImGui.Text(Inline.Span($"Repeating: {(_glfwInput.IsKeyRepeating(Keys.Space) ? "true" : "false")}"));
+			ImGui.Text(Inline.Span($"Pressed: {(_glfwInput.IsKeyPressed(Keys.Space) ? "true" : "false")}"));
+			ImGui.Text(Inline.Span($"Released: {(_glfwInput.IsKeyReleased(Keys.Space) ? "true" : "false")}"));
+
+			ImGui.SeparatorText("LMB state");
+
+			ImGui.Text(Inline.Span($"Down: {(_glfwInput.IsMouseButtonDown(MouseButton.Left) ? "true" : "false")}"));
+			ImGui.Text(Inline.Span($"Pressed: {(_glfwInput.IsMouseButtonPressed(MouseButton.Left) ? "true" : "false")}"));
+			ImGui.Text(Inline.Span($"Released: {(_glfwInput.IsMouseButtonReleased(MouseButton.Left) ? "true" : "false")}"));
+
+			ImGui.SeparatorText("GLFW mouse buttons");
+
+			if (ImGui.BeginTable("GLFW mouse buttons", 8))
+			{
+				for (int i = 0; i < 8; i++)
+				{
+					if (i == 0)
+						ImGui.TableNextRow();
+
+					MouseButton button = (MouseButton)i;
+					if (!Enum.IsDefined(button))
+						continue;
+
+					bool isDown = _glfwInput.IsMouseButtonDown(button);
+
+					ImGui.TableNextColumn();
+
+					if (!_mouseButtonDisplayStringCache.TryGetValue(button, out string? displayString))
+					{
+						displayString = button.ToString();
+						_mouseButtonDisplayStringCache[button] = displayString;
+					}
+
+					ImGui.TextColored(isDown ? Rgba.White : Rgba.Gray(0.4f), displayString);
+				}
+
+				ImGui.EndTable();
+			}
+
+			ImGui.SeparatorText("GLFW mouse wheel");
+
+			ImGui.Text(Inline.Span($"Y: {_glfwInput.MouseWheelY:0.00}"));
+
+			ImGui.SeparatorText("GLFW mouse position");
+
+			ImGui.Text(Inline.Span(_glfwInput.CursorPosition));
 		}
 
 		ImGui.End();